added a bunch of changes that allows either all champions from ea ch generation to be viewed or the best champion overall to be viewed. Also a command line parameter was added that allows for the champions of each generation to be saved or not

+++ src/edu/utexas/cs/nn/tasks/LonerTask.java
@@ -247,19 +247,21 @@ public abstract class LonerTask<T> implements SinglePopulationTask<T> {
         }
 
         if (CommonConstants.netio) {
+        	int currentGen = MMNEAT.ea.currentGeneration();
+        	String filePrefix = "gen" + currentGen + "_";
             // Save best in each objective
             String bestDir = FileUtilities.getSaveDirectory() + "/bestObjectives";
             File dir = new File(bestDir);
             // Delete old contents/team
-            if (dir.exists()) {
+            if (dir.exists() && !Parameters.parameters.booleanParameter("saveAllChampions")) {
                 FileUtilities.deleteDirectoryContents(dir);
             } else {
                 dir.mkdir();
             }
             //save all of the best objectives
             for (int j = 0; j < bestObjectives.length; j++) {
-                Easy.save(bestGenotypes[j], bestDir + "/bestIn" + j + ".xml");
-                FileUtilities.simpleFileWrite(bestDir + "/score" + j + ".txt", bestScores[j].toString());
+                Easy.save(bestGenotypes[j], bestDir + "/" + filePrefix + "bestIn" + j + ".xml");
+                FileUtilities.simpleFileWrite(bestDir + "/" + filePrefix + "score" + j + ".txt", bestScores[j].toString());
             }
         }
